poc/requirements.txt
flask
google-auth
google-auth-oauthlib
google-api-python-client
boxsdk
sentence-transformers
faiss-cpu
sqlalchemy
python-dotenv
requests
python-telegram-bot
